+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix is:hasmod search to include some more mods.
 * You can now drag items into the loadout drawer.
 * D2 spreadsheet export (in settings) covers perks now.
+  * You can also export ghosts (with perks) for D1/D2.
 
 # 4.72.0 (2018-09-30)